src/requirements.txt
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.10" and python_version < "4.0"
beanie==1.24.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cryptography==41.0.7 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.4.2 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.104.1 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.6 ; python_version >= "3.10" and python_version < "4.0"
lazy-model==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
motor==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
orjson==3.9.10 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.14.6 ; python_version >= "3.10" and python_version < "4.0"
pydantic-settings==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.5.3 ; python_version >= "3.10" and python_version < "4.0"
pyjwt[crypto]==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
pymongo==4.6.1 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
python-logstash==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
sentry-sdk==1.39.2 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
structlog==23.3.0 ; python_version >= "3.10" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.24.0.post1 ; python_version >= "3.10" and python_version < "4.0"
